# Conflicts:
#	.github/workflows/e2e-test.yml
#	.github/workflows/mftf-test.yml
#	Api/AdyenPaymentMethodManagementInterface.php
#	Api/AdyenPaymentsDetailsInterface.php
#	Api/Data/InvoiceInterface.php
#	Api/Data/OrderPaymentInterface.php
#	Api/GuestAdyenPaymentMethodManagementInterface.php
#	Api/Internal/InternalAdyenOrderPaymentStatusInterface.php
#	Api/Internal/InternalAdyenPaymentDetailsInterface.php
#	Api/Internal/InternalAdyenPaymentMethodManagementInterface.php
#	Api/Internal/InternalGuestAdyenPaymentMethodManagementInterface.php
#	Block/Form/Moto.php
#	Controller/Return/Index.php
#	Gateway/Http/Client/TransactionPosCloudSync.php
#	Gateway/Request/CheckoutDataBuilder.php
#	Helper/Data.php
#	Helper/PaymentMethods.php
#	Helper/PaymentMethods/ApplePayPaymentMethod.php
#	Helper/PaymentResponseHandler.php
#	Helper/PaymentsDetails.php
#	Helper/Vault.php
#	Model/Api/AdyenDonations.php
#	Model/Api/AdyenOrderPaymentStatus.php
#	Model/Api/AdyenPaymentMethodManagement.php
#	Model/Api/GuestAdyenPaymentMethodManagement.php
#	Model/Api/Internal/InternalAdyenOrderPaymentStatus.php
#	Model/Api/Internal/InternalAdyenPaymentDetails.php
#	Model/Api/Internal/InternalAdyenPaymentMethodManagement.php
#	Model/Api/Internal/InternalGuestAdyenPaymentMethodManagement.php
#	Model/Api/PaymentRequest.php
#	Model/Invoice.php
#	Model/Methods/AmazonPay.php
#	Model/Methods/BcmcMobile.php
#	Model/Methods/Dotpay.php
#	Model/Methods/GooglePay.php
#	Model/Methods/Ideal.php
#	Model/Methods/Klarna.php
#	Model/Methods/Paypal.php
#	Model/Order/Payment.php
#	Model/Resolver/GetAdyenPaymentMethods.php
#	Test/Unit/Helper/InvoiceTest.php
#	composer.json
#	etc/adminhtml/system/adyen_order_processing.xml
#	etc/adminhtml/system/adyen_pwa.xml
#	etc/config.xml
#	etc/di.xml
#	etc/events.xml
#	etc/frontend/di.xml
#	view/frontend/web/js/model/adyen-payment-service.js
#	view/frontend/web/js/view/payment/adyen-methods.js
#	view/frontend/web/js/view/payment/method-renderer/adyen-amazonpay-method.js
#	view/frontend/web/js/view/payment/method-renderer/adyen-bcmc-method.js
#	view/frontend/web/js/view/payment/method-renderer/adyen-dotpay-method.js
#	view/frontend/web/js/view/payment/method-renderer/adyen-googlepay-method.js
#	view/frontend/web/js/view/payment/method-renderer/adyen-ideal-method.js
#	view/frontend/web/js/view/payment/method-renderer/adyen-klarna-method.js
#	view/frontend/web/js/view/payment/method-renderer/adyen-paypal-method.js
#	view/frontend/web/js/view/payment/method-renderer/adyen-pm-method.js
#	view/frontend/web/js/view/payment/method-renderer/vault.js
#	view/frontend/web/template/payment/hpp-form.html
#	view/frontend/web/template/payment/multishipping/hpp-form.html

+++ Api/Data/InvoiceInterface.php
@@ -70,15 +70,15 @@ interface InvoiceInterface
     public function setInvoiceId(int $invoiceId): InvoiceInterface;
 
     /**
-     * @return int|null
+     * @return float|null
      */
-    public function getAmount(): ?int;
+    public function getAmount(): ?float;
 
     /**
-     * @param int $amount
+     * @param float float
      * @return InvoiceInterface
      */
-    public function setAmount(int $amount): InvoiceInterface;
+    public function setAmount(float $amount): InvoiceInterface;
 
     /**
      * @return int|null
